Change of architecture of directories. New subdirectories src/package and include/package
