+++ include/package/types.hpp
@@ -1,0 +1,13 @@
+//
+// Created by eszal on 6.01.2026.
+//
+
+#ifndef NETSIM_TYPES_HPP
+#define NETSIM_TYPES_HPP
+using ElementID = unsigned int;
+
+enum class PackageQueueType {
+    FIFO, LIFO
+};
+
+#endif //NETSIM_TYPES_HPP
